Fixed issue with repo languages, many bugfixes
Fixed issue where the languages for each repository would not appear in the summary.
Changed API call that fetches repos to "user_repositories_url" instead of "repository_search_url" since it is more suited to this website's purposes. The prior method was not capable of fetching repos that were forked by the specified user.
Fixed document title.
Allowed for newlines to be printable on "p" elements for pretty printing.
Added many comments and reordered functions for readability.