Added JSON view/pretty printing
Added the ability to filter if the results should include forks made by the specified user.
If a username does not exist, the summary box will indicate that the user could not be found.
User can also select to view the summary as either JSON or in a rendered format. The summary is lazy loaded to prevent unnecessary API calls.
Fixed issue where the languages were not being fetched with the GitHub authentication header.